Moved SVCSMO to the same file as SVC and made the testing plots a bit more informative..
