+++ Tests/TestingSupportVectorClassification.py
@@ -1,5 +1,4 @@
-from src.DLL.MachineLearning.SupportVectorMachines import SVCSMO
-from src.DLL.MachineLearning.SupportVectorMachines import SVC
+from src.DLL.MachineLearning.SupportVectorMachines.SVC import SVC, SVCSMO
 from src.DLL.MachineLearning import Kernels
 from src.DLL.Data.Preprocessing import data_split
 from src.DLL.Data.Metrics import accuracy
@@ -18,7 +17,7 @@ X = torch.from_numpy(X)
 y = torch.from_numpy(y)
 X_train, y_train, X_test, y_test, _, _ = data_split(X, y)
 
-model1 = SVC.SVC(kernel=Kernels.SquaredExponentialCovariance())
+model1 = SVC(kernel=Kernels.LinearCovariance())
 model1.fit(X_train, y_train)
 predictions = model1.predict(X_test)
 print(round(accuracy(predictions, y_test), 3))
@@ -33,7 +32,7 @@ Xv, Yv = torch.meshgrid(x_grid, y_grid, indexing="ij")
 X_grid = torch.stack((Xv.flatten(), Yv.flatten()), dim=1).to(X.dtype)
 grid_predictions = model1.predict(X_grid).reshape((n, n))
 
-model2 = SVCSMO.SVC(kernel=Kernels.SquaredExponentialCovariance())
+model2 = SVCSMO(kernel=Kernels.LinearCovariance())
 model2.fit(X_train, y_train, epochs=1000)
 predictions = model2.predict(X_test)
 print(round(accuracy(predictions, y_test), 3))
@@ -53,4 +52,6 @@ axes[0].contourf(Xv.numpy(), Yv.numpy(), grid_predictions.numpy(), alpha=0.5)
 axes[1].contourf(Xv.numpy(), Yv.numpy(), grid_predictions2.numpy(), alpha=0.5)
 axes[0].scatter(X[:, 0], X[:, 1], c=y, s=5)
 axes[1].scatter(X[:, 0], X[:, 1], c=y, s=5)
+axes[0].set_title("cvxopt solver")
+axes[1].set_title("My SMO solver")
 plt.show()
